Manage files with Stow and clean up Bash files
Managing dotfiles with Stow instead of setup scripts. This required
restructuring the directories.

For more info on stow:
https://www.gnu.org/software/stow/

Noticed a lot of cruft--duplicate aliases and file sourcing in
`bash_profile` and `bash_aliases`--while moving files around for stow.

Also, deleted Terminator, as it had not been used in quite some time.