Make Space 2 floating with no borders
Since space two is used for applications that do not play nice with
Kwm, it has been set to always be in `float` mode.

To remove the window borders in this space, a small script checks which
space is currently active when a switch occurs.

There are definitely some kinks to work out, but hopefully this
improves osme of the annoyances I have been having when trying to use
applications in this space.