Automate creation of dated design files
_Today's Working File_, or `twf`, copies a file and changes the nested
date in the filename to today. It also has an option to change the
filename with a second argument.

This should help stop the situation where I would open a project
binary file and forget to "Save As", overwriting previous work without
the ability to recover it.

Hopeful this is a stop gap, and I will be able to change my workflow to
include `git_lfs`.