Nvim: fix auto-diagnostic clashing with LSP hover
The auto diagnostic on cursor hover would not allow LSP hover or
signature help to open. The fix involved moving the handlers to the
default LSP config and sharing a global `diagnostic_hover` boolean that
gets turned off diagnostic hover when those two actions are executed and
then re-enabled when the float window is closed.