Zsh: fix audio drift when speeding up recording
Adds `async` flag to ffmpeg in the `mov2mp4` alias. Sometimes when
converting a longer recording to 1.4x, the audio and video would start
to drift. Supposedly this flag fixes this, but the jury is still out.
Noticed some drift when using an uneven speedup of 1.3, but 1.2 and 1.4
seem to work.